Key error in data structure
I didn't look into this in detail, but the previous version didn't match the data structure.

P.s. Could you please also make releases of this repo, or at least tag versions every now and then? It's difficult to integrate this in a workflow when you cannot pin versions. Thanks!